+++ src/idtrackerai/session.py
@@ -563,7 +563,10 @@ class Session:
     @property
     def bkg_model(self) -> np.ndarray | None:
         if self.background_path.is_file():
-            return cv2.imread(str(self.background_path))[..., 0]
+            # Use fromfile to handle paths with non-ASCII characters
+            return cv2.imdecode(
+                np.fromfile(self.background_path, dtype=np.uint8), cv2.IMREAD_COLOR_BGR
+            )[..., 0]
         return None
 
     @bkg_model.setter
@@ -582,7 +585,9 @@ class Session:
     @property
     def ROI_mask(self) -> np.ndarray | None:
         if self.ROI_mask_path.is_file():
-            return cv2.imread(str(self.ROI_mask_path))[..., 0]
+            return cv2.imdecode(
+                np.fromfile(self.ROI_mask_path, dtype=np.uint8), cv2.IMREAD_COLOR_BGR
+            )[..., 0]
         return None
 
     @ROI_mask.setter